feat(Error): add IOException functions (bump 2.1.0.0)
These are helpers for working with exceptions, but still encouraging a
“pure” handling of error values.

It is a bit inefficient to decode back and forth from strings, thus
`ifIOError` will throw an opaque exception type, which cannot be
caught again, to discourage exception-heavy code.

Basically we want to make it easy to handle library code that is
already throwing, and convert it to Errors which can be annotated for
fun & profit (& hopefully good error messages).